# perf(thrift): optimized skip decoder